feat: use context for markets
